+++ modules/crypto/test/index.js
@@ -1,0 +1,3 @@
+import './lib/crypto-hash-transform.spec';
+import './lib/crc32-hash-transform.spec';
+import './lib/md5-wasm-transform.spec';