+++ modules/loader-utils/src/lib/iterator-utils/make-transform-iterator.d.ts
@@ -1,0 +1,13 @@
+import {IncrementalTransform} from './incremental-transform';
+
+/**
+ *
+ * @param asyncIterator
+ * @param IncrementalTransform
+ * @param options
+ */
+export function makeTransformIterator(
+  asyncIterator: AsyncIterable<ArrayBuffer> | Iterable<ArrayBuffer>,
+  transform: IncrementalTransform,
+  options?: object
+): AsyncIterable<any>;