+++ modules/loader-utils/test/index.js
@@ -4,6 +4,7 @@ import './lib/binary-utils/memory-copy-utils.spec';
 
 import './lib/iterator-utils/async-iteration.spec';
 import './lib/iterator-utils/make-iterator.spec';
+import './lib/iterator-utils/make-transform-iterator.spec';
 
 import './lib/library-utils/require-utils.spec';
 import './lib/library-utils/library-utils.spec';